+++ src/part/reactor.h
@@ -5,9 +5,9 @@ class reactor : protected part {
 	public:
 		reactor();
 		void run();
-		bool adjustOutput( const float &fnewpercent, const bool &overloadenabled );
+		bool adjustOutput( const double &fnewpercent, const bool &overloadenabled );
 	protected:
-		float p_coreUtilization;
+		double p_coreUtilization;
 		int p_maxPGridMod;
 		int p_coreTemp;
 		bool p_isRunning;
